requirements.txt
numpy==1.26.4
scikit-learn==1.7.0
scikit-image==0.22.0
diffusers==0.34.0
transformers==4.53.2
accelerate==1.9.0
einops==0.8.1
pillow==11.0.0
tqdm==4.67.1
porespy==2.4.2
openpnm==3.5.0
matplotlib==3.8.4
h5py==3.11.0
safetensors==0.5.3